+++ public/post/exeDelete.php
@@ -1,7 +1,10 @@
 <?php
 
-$entity = filter_input(INPUT_POST, 'entity', FILTER_DEFAULT);
+$entity = trim(strip_tags(filter_input(INPUT_POST, 'entity')));
 $id = filter_input(INPUT_POST, 'id', FILTER_VALIDATE_INT);
 
-$result = \Entity\Entity::delete($entity, $id);
-$data['data'] = $result;
+$data['data'] = "";
+if(!empty($entity) && !empty($id)) {
+    $result = \Entity\Entity::delete($entity, $id);
+    $data['data'] = $result;
+}
